requirements.txt
torch
torchvision
facenet-pytorch
streamlit
opencv-python
numpy
matplotlib
scikit-learn
tqdm
Pillow